# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText OpenAPI
  description: |-
    This OpenAPI document describes the WasaText API.
    Note: One-on-one conversations are implemented as "Groups" with 2 person.
    To distinguish between group conversations (also 2 people),
    the enum field "Conversation Type" is used.
  version: "1"
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: "User details"
        content:
          application/json:
            schema:
              description: "The Username, which is a unique identifier"
              type: object
              properties:
                Username: 
                  type: string
                  description: "a unique identifier of the user"
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: "User log-in action successful"
          content:
            application/json:
              schema:
                description: "The login identifier"
                type: object
                properties:
                  identifier:
                    # change here if you decide to use an integer
                    # or any other type of identifier
                    type: string
                    example: "Maria"
                    pattern: "ˆ.*$"
                    minLength: 1
                    maxLength: 100
...
  /conversation/{ConversationID}:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
    get:
      tags: ["Conversation Management"]
      summary: get the selected conversation
      description: "Get the selected conversations"
      operationId: getConversation
      responses:
        "200":
          description: "The selected conversation was fetched successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
           description: "User or Conversation not found"
    post:
      tags: ["Conversation Management"]
      summary: send a message to the conversation
      description: "The user sends a message to the conversation"
      operationId: sendMessage
      requestBody:
        description: "The message to send to the conversation"
        content:
          application/json:
            schema:
              description: "The content of the message"
              type: object
              properties:
                MessageContent:
                  type: string
                  description: "the content of the message"
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 1
                  maxLength: 10000
        required: true
      responses:
        "200":
          description: "The message to the conversation was sent successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
           description: "User or Conversation not found"
    put:
      tags: ["Conversation Management"]
      summary: set the group name
      description: "Set the group name"
      operationId: setGroupName
      requestBody:
        description: "Group Details"
        content:
          application/json:
            schema:
              description: "The new GroupName"
              type: object
              properties:
                GroupName: 
                  type: string
                  description: "new name for the group"
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "200":
          description: "Group Details set successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "403":
          description: "Forbidden request, cannot set group name of one-on-one conversation"
        "404":
          description: "User or Conversation not found"
          
    delete:
      tags: ["Conversation Management"]
      summary: leave group
      description: "Leave from the group"
      operationId: leaveGroup
      responses:
        "200":
          description: "Left from the group successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyConversations"
        "404":
           description: "User or Conversation not found"
  /conversation/{ConversationID}/add:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
    put:
      tags: ["Conversation Management"]
      summary: Add Someone To Group
      description: "Add the specified user to the group"
      operationId: addToGroup
      requestBody:
        description: "User to Add"
        content:
          application/json:
            schema:
              description: "The user to add to the group"
              type: object
              properties:
                UserNameToAdd: 
                  type: string
                  description: "User to add to the group"
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "200":
          description: "User has been added to the Group successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
           description: "User or Conversation not found"
  /conversation/{ConversationID}/message/{MessageID}:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
      - $ref: "#/components/parameters/MessageID"
    delete:
      tags: ["Message Management"]
      summary: delete message
      description: "delete the message from the conversation"
      operationId: deleteMessage
      responses:
        "200":
          description: "Message deleted successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
           description: "User, Conversation or Message not found"
  /conversation/{ConversationID}/message/{MessageID}/comment:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
      - $ref: "#/components/parameters/MessageID"
    post:
      tags: ["Message Management"]
      summary: Forward a Message
      description: "Forward the message to someone"
      operationId: forwardMessage
      requestBody:
        description: "Forward The Message"
        content:
          application/json:
            schema:
              description: "The recipient Username"
              type: object
              properties:
                RecipientUsername:
                  description: "The recipient of the message to be forwarded"
                  type: string
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 3
                  maxLength: 16
      responses:
        "200":
            description: "Message forwarded successfully"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Message"
        "404": 
          description: "User, Conversation or Message not found"
    put:
      tags: ["Message Management"]
      summary: React to Message
      description: "Send a reaction or comment to a message"
      operationId: commentMessage
      requestBody:
        description: "Comment The Message"
        content:
          application/json:
            schema:
              description: "The type of reaction and the content of the reaction"
              type: object
              properties:
                ReactionType:
                  description: "the type of reaction, can be either emoji or message"
                  type: string
                  enum:
                    - EmojiReaction
                    - MessageReaction
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 1
                  maxLength: 7
                Content: 
                  type: string
                  description: "The content of the reaction; if the reaction is a message |
                                this is a reply and so the content is the message content; |
                                If the reaction is an emoji, this is the encoding of the emoji"
                  example: "Maria"
                  pattern: "ˆ.*$"
                  minLength: 1
                  maxLength: 10000
        required: true
      responses:
        "200":
            description: "Message reaction sent successfully"
            content:
              application/json:
                schema: 
                  $ref: "#/components/schemas/Message"
        "404": 
          description: "User, Conversation or Message not found"
    delete:
      tags: ["Message Management"]
      summary: uncomment message
      description: "delete the comment or reaction from the message"
      operationId: uncommentMessage
      responses:
        "200":
          description: "Message reaction deleted successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
           description: "User, Conversation or Message not found"
...
components:
  schemas:
    User:
      description: "Object representing a User"
      type: object
      properties:
        Username:
          description: "a unique identifier of the user"
          type: string
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 16
        Profile Picture:
          description: "the profile picture of the user"
          type: string
          format: binary
          minLength: 1
          maxLength: 100000
    Message:
      description: "Object representing a single message"
      type: object
      properties:
        Message ID:
          description: "message id that is unique in the scope of a conversation"
          type: integer
        Sender:
          description: "the user of the person who sent the message"
          allOf:
            - $ref: "#/components/schemas/User"
            - description: "the user of the person who sent the message"
        Content:
          description: "the content of the message"
          type: string
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1 #the message has to be at least one character
          maxLength: 2000 #limiting the max message length
          
        Timestamp:
          description: "the timestamp of the message"
          type: string
          format: date-time
          minLength: 1
          maxLength: 100
          
        Status:
          description: "the status of the message"
          type: string
          enum:
            - SingleCheckmark
            - DoubleCheckmark
            - Username
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 10000
        Reactions:
          description: "all reactions to this message"
          type: array
          items:
            $ref: "#/components/schemas/Reaction" 
          minItems: 0
          maxItems: 10000
    Reaction:
      description: "Object representing a reaction to a message"
      type: object
      properties:
        UserWhoReacted:
          description: "The user who made the reaction"
          allOf:
            - $ref: "#/components/schemas/User"
            - description: "The user who made the reaction"
        ReactionType:
          description: "the type of reaction, can be either emoji or message"
          type: string
          enum:
            - EmojiReaction
            - MessageReaction
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 10000
        Content:
          description: "the content of reaction, can be either a message or the encoding of an emoji"
          type: string
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 10000
    Conversation:
      description: "Object representing a conversation"
      type: object
      properties:
        id:
          description: "identifier of a conversation, which is either
            the unique username of the other person or
            the unique Group ID of the group"
          type: string
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 10000
        Participants:
          description: "users in the group"
          type: array
          items: 
            $ref: "#/components/schemas/User"
          minItems: 2
          maxItems: 10000
        Group name:
          description: "name of the group"
          type: string
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 3
          maxLength: 20
        Group Picture:
          description: "the picture of the group"
          type: string
          format: binary
          minLength: 1
          maxLength: 100
        ConversationType:
          description: "type of conversation, either with a single person or a group"
          type: string
          enum:
            - UserType
            - GroupType
          example: "Maria"
          pattern: "ˆ.*$"
          minLength: 1
          maxLength: 5
        Messages:
          description: "content of the conversation -> a list of messages"
          type: array
          items: 
            $ref: "#/components/schemas/Message"
          minItems: 1
          maxItems: 10000
    MyConversations:
      description: "Object storing all of my conversations"
      type: array
      items:
        $ref: "#/components/schemas/Conversation"
      minItems: 0
      maxItems: 10000
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      description: "Bearer authentication"
      type: http
      scheme: bearer
  parameters:
    ConversationID:
      description: "Unique ID for the conversation"
      schema:
        description: "Unique ID for the conversation"
        type: string
        example: "Maria"
        pattern: "ˆ.*$"
        minLength: 1
        maxLength: 10000
        readOnly: true
      name: ConversationID
      in: path
      required: true
    MessageID:
      description: "Message ID that is unique in the scope of a conversation"
      schema:
        description: "Message ID that is unique in the scope of a conversation"
        type: integer
        readOnly: true
      name: MessageID
      in: path
      required: true
security:
  - bearerAuth: []